Clean up everything (see description)
Extract tooltip lines & axes to components, simplify / remove repeated logic, add utils file
