+++ components/dashboard/display/placeholder/index.tsx
